refactor : profile header widght
