+++ frontend-hybrid/yakal/lib/screens/Home/home_screen.dart
@@ -6,15 +6,10 @@ class HomeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Center(
+    return const Center(
       child: Column(
         children: [
-          const Text('홈'),
-          SvgPicture.asset(
-            'assets/icons/icon-health.svg',
-            width: 160,
-            height: 160,
-          ),
+          Text('홈'),
         ],
       ),
     );
