+++ frontend-hybrid/yakal/lib/screens/Profile/profile_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:yakal/viewModels/Profile/user_view_model.dart';
-import 'package:yakal/widgets/Profile/profie_header_widget.dart'; // Import the AuthLoginScreen
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:yakal/widgets/Profile/ProfileHeader/profile_header_widget.dart';
 // 임시 유저 클래스
 
 class ProfileScreen extends StatelessWidget {
@@ -17,57 +16,9 @@ class ProfileScreen extends StatelessWidget {
       body: Column(
         mainAxisAlignment: MainAxisAlignment.start,
         children: [
-          Container(
-            color: Colors.white,
-            // height 240px
-            height: 260,
+          // 프로필 헤더
+          ProfileHeader(userViewModel: userViewModel),
 
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  const SizedBox(height: 80),
-                  ProfileHeader(userViewModel: userViewModel),
-                  const SizedBox(height: 30),
-                  OutlinedButton(
-                    // backgorund : rgba(233, 233, 238, 1)
-                    style: OutlinedButton.styleFrom(
-                      padding: const EdgeInsets.symmetric(
-                          vertical: 8, horizontal: 12), // Adjust padding
-                      foregroundColor: Colors.black,
-                      backgroundColor: const Color.fromRGBO(233, 233, 238, 1),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      // border none
-                      side: const BorderSide(
-                        color: Color.fromRGBO(233, 233, 238, 1),
-                      ),
-                    ),
-
-                    onPressed: () {
-                      // Get.toNamed("/login");
-                    },
-                    child: const Row(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: [
-                        // SvgPicture.asset(
-                        //   'assets/images/icon-health.svg',
-                        //   width: 20,
-                        //   height: 20,
-                        // ),
-                        Icon(Icons.edit),
-                        Text('닉네임 수정'),
-                        SizedBox(width: 10),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            // ROW BACCK GORUND WHITE
-          ),
           Obx(() {
             return Text('Nickname: ${userViewModel.user.value.nickName}');
           }),
